backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymupdf>=1.24.0
python-docx==1.1.0
sentence-transformers>=2.2.2
scikit-learn>=1.4.0
numpy>=1.26.0
requests==2.31.0
reportlab>=4.0.0

